Handlers and batch size complete, TODO: prime algo, main exit checking, signal handling, argv expansion

+++ include/utils/utils.hpp
@@ -1,15 +1,15 @@
-int** distribute_range(int lower_bound, int upper_bound, int elements) {
-  int** boundaries = new int*[elements];
+long int** distribute_range(long int lower_bound, long int upper_bound, int elements) {
+  long int** boundaries = new long int*[elements];
 
   int distribution_range = (upper_bound - lower_bound) / elements;
   int range_remainder = (upper_bound - lower_bound) % elements;
   int remainders_added = 0;
 
   for (int i = 0; i < elements; i++) {
-    boundaries[i] = new int[2];
+    boundaries[i] = new long int[2];
 
-    int element_lower_bound = lower_bound + remainders_added + (i*distribution_range);
-    int element_upper_bound = element_lower_bound + distribution_range - 1;
+    long int element_lower_bound = lower_bound + remainders_added + (i*distribution_range);
+    long int element_upper_bound = element_lower_bound + distribution_range - 1;
 
     if (remainders_added <= range_remainder) {
       element_upper_bound++;
